requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.24.36
botocore==1.27.36
Django==4.0.6
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-jalali==5.0.0
django-jalali-date==1.0.1
django-jquery-js==3.1.1
django-storages==1.12.3
jdatetime==3.8.0
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.10
